+++ x/wasm/keeper/keeper_test.go
@@ -390,7 +390,7 @@ func TestInstantiate(t *testing.T) {
 
 	gasAfter := ctx.GasMeter().GasConsumed()
 	if types.EnableGasVerification {
-		require.Equal(t, uint64(0x18bfc), gasAfter-gasBefore)
+		require.Equal(t, uint64(0x18c06), gasAfter-gasBefore)
 	}
 
 	// ensure it is stored properly